+++ src/App.js
@@ -1,5 +1,5 @@
 import Title from './components/pure/Title.jsx';
-import Form from './components/containers/Form.jsx';
+import Form from './components/forms/Form.jsx';
 
 function App() {
   return (
